+++ src/containers/forms/PlanForm/index.tsx
@@ -16,6 +16,7 @@ import {
 } from 'reactstrap';
 import { format } from 'util';
 import DatePickerWrapper from '../../../components/DatePickerWrapper';
+import JurisdictionSelect from '../../../components/forms/JurisdictionSelect';
 import {
   DATE_FORMAT,
   DEFAULT_PLAN_DURATION_DAYS,
@@ -66,7 +67,6 @@ import {
 import { PLAN_LIST_URL } from '../../../constants';
 import { OpenSRPService } from '../../../services/opensrp';
 import { InterventionType, PlanStatus } from '../../../store/ducks/plans';
-import JurisdictionSelect from '../JurisdictionSelect';
 import {
   doesFieldHaveErrors,
   FIActivities,
